WEB: Move Game Demos to ORM
