+++ include/Models/GameDemosModel.php
@@ -1,38 +1,24 @@
 <?php
 namespace ScummVM\Models;
 
-use ScummVM\Objects\GameDemo;
-use ScummVM\Models\GameModel;
-use Symfony\Component\Yaml\Yaml;
+use ScummVM\OrmObjects\DemoQuery;
 
 /**
  * The GameDemosModel class will generate GameDemo objects.
  */
 class GameDemosModel extends BasicModel
 {
-    private $gameModel;
-    private $platformsModel;
-
-    public function __construct()
-    {
-        $this->gameModel = new GameModel();
-        $this->platformsModel = new SimpleModel("Platform", "platforms.yaml");
-    }
     /* Get all the groups and their respective demos. */
     public function getAllGroupsAndDemos()
     {
         $groupedData = $this->getFromCache();
         if (is_null($groupedData)) {
-            $fname = $this->getLocalizedFile('game_demos.yaml');
-            $demos = Yaml::parseFile($fname);
-            $games = $this->gameModel->getAllGames();
-            $platforms = $this->platformsModel->getAllData();
-            $data = [];
-            foreach ($demos as $demo) {
-                $obj = new GameDemo($demo, $games, $platforms);
-                $data[] = $obj;
-            }
-            $groupedData =  $this->createGroups($data);
+            $demos = DemoQuery::create()
+                ->useGameQuery()
+                ->orderByName()
+                ->endUse()
+            ->find();
+            $groupedData =  $this->createGroups($demos);
             $this->saveToCache($groupedData);
         }
         return $groupedData;
@@ -43,13 +29,14 @@ class GameDemosModel extends BasicModel
         $groups = [];
         foreach ($demos as $demo) {
             $company = $demo->getGame()->getCompany();
-            if (is_string($company)) {
+
+            if ($company === null) {
                 $companyName = "Unknown";
             } else {
                 $companyName = $company->getName();
             }
             $companyId = $company->getId();
-            if (!array_key_exists($companyId, $groups)) {
+            if (!isset($groups[$companyId])) {
                 $groups[$companyId] = [
                     'name' => "$companyName Demos",
                     'href' => $companyId,
@@ -67,7 +54,6 @@ class GameDemosModel extends BasicModel
             'demos' => []
         ];
         foreach ($groups as $key => $group) {
-            \sort($groups[$key]['demos'], SORT_STRING);
             if (count($groups[$key]['demos']) <= 15) {
                 $groups['other']['demos'] = \array_merge($groups['other']['demos'], $groups[$key]['demos']);
                 unset($groups[$key]);
